backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coreapi>=2.3.3",
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
]
